pyproject.toml
[tool.poetry]
name = "Dimension"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
langchain-community = "^0.2.12"
langchain-openai = "^0.1.22"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0.post1"
langchain-experimental = "^0.0.64"
aiofiles = "^24.1.0"
pyarrow = "^17.0.0"
cachetools = "^5.5.0"
pandas = "^2.2.2"
python-multipart = "^0.0.9"
tabulate = "^0.9.0"
chardet = "^5.2.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
